Fix: troca da pasta tabuleiro por maquina de regras
