+++ config/SetupPecas.java
@@ -3,6 +3,7 @@ package config;
 import java.util.ArrayList;
 import java.util.HashMap;
 
+import maquinaDeRegras.Tabuleiro;
 import pecas.Bispo;
 import pecas.Cavalo;
 import pecas.Dama;
@@ -11,7 +12,6 @@ import pecas.Peca;
 import pecas.Rei;
 import pecas.TipoPeca;
 import pecas.Torre;
-import tabuleiro.Tabuleiro;
 import utils.Cor;
 import utils.Posicao;
